Switch to Full Results API with DDG fallback for images
This removes the use of the WolframAlpha Python package instead
querying the API's directly. If a personal API Key is supplied, the
API will be queried directly, otherwise the Mycroft backend will be
used.

For images that are expected to be text only we do not look for an
image. Otherwise we do our best to get an image from the Wolfram
response, and fallback to a DuckDuckGo image search.

This includes a fairly significant refactor of the existing Skill.